Completely revamped the GameBadge. Made the interface mobile friendly. The style has been changed for better readability and use on mobile/desktop applications. Proper ellipsis will appear when descriptions/names don’t fit in allocated spaces.
Added expandable “full details” panel where users can download the client log as well as read the full description.

Removed the grid above the badges as it was a major issue with resizability.
Added lower bound to size at 320px to avoid ugly formatting.

Added Ellipsis and Collapse class attributes as well as gutter shortening for the Bootstrap grid to main.css